split object-inserter, deleter and node analyzer into seperate files
