+++ include/oos/object/object_store.hpp
@@ -943,262 +943,18 @@ object_store::iterator object_store::initialize(prototype_node *node, const ON_A
 
 namespace detail {
 
-template<class T, template < class ... > class ON_ATTACH>
-template < class V >
-void node_analyzer<T, ON_ATTACH>::serialize(V &x)
-{
-  oos::access::serialize(*this, x);
-}
-
-template<class T, template < class ... > class ON_ATTACH>
-template<class V>
-void node_analyzer<T, ON_ATTACH>::serialize(const char *id, belongs_to<V> &, cascade_type)
-{
-  node_.register_belongs_to(std::type_index(typeid(V)), id);
-}
-
-template<class T, template < class ... > class ON_ATTACH>
-template<class V>
-void node_analyzer<T, ON_ATTACH>::serialize(const char *, has_one<V> &x, cascade_type)
-{
-  prototype_iterator node = node_.tree()->find(x.type());
-  if (node == node_.tree()->end()) {
-    // if there is no such prototype node
-    // prepare insertion of new node
-    node = node_.tree()->template prepare_attach<V>();
-  } else if (!node->has_primary_key()) {
-    throw_object_exception("serializable of type '" << x.type() << "' has no primary key");
-  }
-}
-
-template<class T, template < class ... > class ON_ATTACH>
-template<class V, template<class ...> class C>
-void node_analyzer<T, ON_ATTACH>::serialize(const char *id, has_many<V, C> &, const char *, const char *)
-//void node_analyzer<T, ON_ATTACH>::serialize(const char *id, has_many<V, C> &x, const char */*owner_field*/, const char */*item_field*/)
-{
-  // item column column names
-//  x.owner_field(owner_field);
-//  x.item_field(item_field);
-  // attach relation table for has many relation
-  // check if has many item is already attached
-  // true: check owner and item field
-  // false: attach it
-  prototype_iterator pi = node_.tree()->find(id);
-  if (pi == node_.tree()->end()) {
-    pi = node_.tree()->template attach<typename has_many<V, C>::item_type, ON_ATTACH>(id, false, nullptr, on_attach_);
-  } else if (pi->type_index() == std::type_index(typeid(typename has_many<V, C>::item_type))) {
-    // prototype is of type has_many_item
-    throw_object_exception("many to many relations are not supported by now");
-  } else {
-    // throw exception
-    throw_object_exception("prototype already inserted: " << pi->type());
-  }
-}
-
 template < class T >
 void modified_marker::marker_func(object_store &store, object_proxy &proxy)
 {
   store.mark_modified<T>(&proxy);
 }
 
-template<class T>
-void object_inserter::insert(object_proxy *proxy, T *o, bool notify)
-{
-
-  notify_ = notify;
-
-  object_proxy_stack_.push(proxy);
-
-  modified_marker_ = [](object_store &store, object_proxy *oproxy) {
-    store.mark_modified<T>(oproxy);
-  };
-
-  if (proxy->obj()) {
-    oos::access::serialize(*this, *o);
-  }
-  object_proxy_stack_.pop();
-}
-
-template<class T>
-void object_inserter::serialize(T &x)
-{
-  oos::access::serialize(*this, x);
-}
-
-template<class T>
-void object_inserter::serialize(const char *, belongs_to<T> &x, cascade_type cascade) {
-  if (x.is_inserted() || (x.proxy_ && x.proxy_->obj() == nullptr)) {
-    return;
-  }
-  x.is_inserted_ = true;
-  x.cascade_ = cascade;
-  // object was seen by inserter stop inserting
-  if (!object_proxies_.insert(x.proxy_).second) {
-    return;
-  }
-
-  if (!x.proxy_) {
-    return;
-  }
-
-  if (x.id()) {
-    // do the pointer count
-    object_proxy_stack_.push(x.proxy_);
-    oos::access::serialize(*this, *(T*)x.ptr());
-    object_proxy_stack_.pop();
-  } else {
-    // new object
-    ostore_.insert<T>(x.proxy_, notify_);
-  }
-  ++(*x.proxy_);
-}
-
-template<class T>
-void object_inserter::serialize(const char *, has_one<T> &x, cascade_type cascade) {
-  if (x.is_inserted() || (x.proxy_ && x.proxy_->obj() == nullptr)) {
-    return;
-  }
-  x.is_inserted_ = true;
-  x.cascade_ = cascade;
-  // object was seen by inserter stop inserting
-  if (!object_proxies_.insert(x.proxy_).second) {
-    return;
-  }
-
-  if (!x.proxy_) {
-    return;
-  }
-
-  if (x.id()) {
-    // do the pointer count
-    object_proxy_stack_.push(x.proxy_);
-    oos::access::serialize(*this, *(T*)x.ptr());
-    object_proxy_stack_.pop();
-  } else {
-    // new object
-    ostore_.insert<T>(x.proxy_, notify_);
-  }
-  ++(*x.proxy_);
-}
-
-template<class T, template<class ...> class C>
-void object_inserter::serialize(const char *, basic_has_many<T, C> &x, const char*, const char*)
-{
-  // initialize the has many relation
-  // set identifier
-  // relation table name
-  // owner column name
-  // item column name
-  if (object_proxy_stack_.empty()) {
-    throw object_exception("no owner for has many relation");
-  }
-
-  if (x.ostore_) {
-    return;
-  }
-  object_proxy *proxy = object_proxy_stack_.top();
-  x.owner_id_ = proxy->pk();
-  x.owner_ = proxy;
-  x.ostore_ = &ostore_;
-  x.mark_modified_owner_ = modified_marker_;
-
-  typename basic_has_many<T, C>::iterator first = x.begin();
-  typename basic_has_many<T, C>::iterator last = x.end();
-
-  while (first != last) {
-    typename basic_has_many<T, C>::relation_type i = (first++).relation_item();
-    if (!i.is_inserted()) {
-      // item is not in store, insert it
-      ostore_.insert(i, false);
-    }
-  }
-}
-
-template <typename T>
-void object_deleter::t_object_count::remove_object(object_proxy *proxy, bool notify)
-{
-  proxy->ostore()->remove<T>(proxy, notify, false);
-}
-
-template<class T>
-bool object_deleter::is_deletable(object_proxy *proxy, T *o) {
-  object_count_map.clear();
-  object_count_map.insert(std::make_pair(proxy->id(), t_object_count(proxy, false, (T*)proxy->obj())));
-
-  // start collecting information
-  oos::access::serialize(*this, *o);
-
-  return check_object_count_map();
-}
-
-template<class T>
-void object_deleter::serialize(const char *, belongs_to<T> &x, cascade_type cascade) {
-  if (!x.ptr()) {
-    return;
-  }
-  std::pair<t_object_count_map::iterator, bool> ret = object_count_map.insert(
-    std::make_pair(x.proxy_->id(), t_object_count(x.proxy_, true, (T*)x.proxy_->obj()))
-  );
-  --ret.first->second.reference_counter;
-  if (cascade & cascade_type::REMOVE) {
-    ret.first->second.ignore = false;
-    oos::access::serialize(*this, *(T*)x.ptr());
-  }
-}
-
-template<class T>
-void object_deleter::serialize(const char *, has_one<T> &x, cascade_type cascade) {
-  if (!x.ptr()) {
-    return;
-  }
-  std::pair<t_object_count_map::iterator, bool> ret = object_count_map.insert(
-    std::make_pair(x.proxy_->id(), t_object_count(x.proxy_, true, (T*)x.proxy_->obj()))
-  );
-  --ret.first->second.reference_counter;
-  if (cascade & cascade_type::REMOVE) {
-    ret.first->second.ignore = false;
-    oos::access::serialize(*this, *(T*)x.ptr());
-  }
-}
-
-template<class T, template<class ...> class C>
-void object_deleter::serialize(const char *, basic_has_many<T, C> &x, const char *, const char *)
-{
-  typename basic_has_many<T, C>::iterator first = x.begin();
-  typename basic_has_many<T, C>::iterator last = x.end();
-  while (first != last) {
-    // Todo: get the real holder: on join table get has_many_item
-    typename basic_has_many<T, C>::relation_type iptr = first.relation_item();
-    ++first;
-    object_proxy *proxy = iptr.proxy_;
-    std::pair<t_object_count_map::iterator, bool> ret = object_count_map.insert(
-      std::make_pair(proxy->id(), t_object_count(proxy, false, (T*)proxy->obj()))
-    );
-    /**********
-     *
-     * object is already in list and will
-     * be ignored on deletion so set
-     * ignore flag to false because this
-     * node must be deleted
-     *
-     **********/
-    if (!ret.second && ret.first->second.ignore) {
-      ret.first->second.ignore = false;
-    }
-
-    oos::access::serialize(*this, *iptr);
-  }
-}
-
-template<class T>
-void object_deleter::serialize(const char *id, identifier<T> &x)
-{
-  T val = x.value();
-  serialize(id, val);
 }
 
 }
 
-}
+#include "oos/object/node_analyzer.tpp"
+#include "oos/object/object_inserter.tpp"
+#include "oos/object/object_deleter.tpp"
 
 #endif /* OBJECT_STORE_HPP */
