+++ src/object/CMakeLists.txt
@@ -14,6 +14,8 @@ SET(SOURCES
   delete_action.cpp
   basic_has_many_item.cpp
   object_proxy_accessor.cpp
+  object_deleter.cpp
+  object_inserter.cpp
 )
 
 SET(HEADER
@@ -50,7 +52,16 @@ SET(HEADER
   ${CMAKE_SOURCE_DIR}/include/oos/object/has_many_item.hpp
   ${CMAKE_SOURCE_DIR}/include/oos/object/basic_has_many_item.hpp
   ${CMAKE_SOURCE_DIR}/include/oos/object/identifier_proxy_map.hpp
-  ${CMAKE_SOURCE_DIR}/include/oos/object/object_proxy_accessor.hpp ../../include/oos/object/belongs_to.hpp ../../include/oos/object/object_holder_type.hpp ../../include/oos/object/object_inserter.hpp ../../include/oos/object/object_deleter.hpp ../../include/oos/object/node_analyzer.hpp)
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_proxy_accessor.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/belongs_to.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_holder_type.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_inserter.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_inserter.tpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_deleter.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/object_deleter.tpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/node_analyzer.hpp
+  ${CMAKE_SOURCE_DIR}/include/oos/object/node_analyzer.tpp
+)
 
 ADD_LIBRARY(oos-object SHARED ${SOURCES} ${HEADER})
